Allow user to create a ticket hold (#442)
* Create hold

* Create route for creating hold

* Add update holds

* Try to speed up cargo downloads

* Added more hold fields and an update method

* Added route

* Add fields to create hold test

* Add hold items to initial hold create

* Add holds tests

+++ api/src/controllers/cart.rs
@@ -23,7 +23,7 @@ pub struct CartResponse {
 #[derive(Deserialize)]
 pub struct AddToCartRequestItem {
     pub ticket_type_id: Uuid,
-    pub quantity: i64,
+    pub quantity: u32,
 }
 
 #[derive(Deserialize)]
@@ -66,7 +66,7 @@ pub fn add(
 #[derive(Deserialize)]
 pub struct RemoveCartRequest {
     pub cart_item_id: Uuid,
-    pub quantity: Option<i64>,
+    pub quantity: Option<u32>,
 }
 
 pub fn remove(
